requirements.txt
beautifulsoup4
couchbase
python-dotenv
huggingface-hub
requests
scikit-learn
sentence-transformers
streamlit
tokenizers
torch
transformers
nest-asyncio
asyncio